+++ source/qip/package.py
@@ -93,7 +93,7 @@ def install(request, path, context_mapping, cache_path, editable_mode=False):
         "--disable-pip-version-check "
         "--cache-dir {cache_dir} "
         "{editable_mode}" 
-        "{requirement}".format(
+        "'{requirement}'".format(
             editable_mode="-e " if editable_mode else "",
             destination=path,
             requirement=request,