+++ doc/release/release_notes.rst
@@ -4,7 +4,8 @@
 Release Notes
 *************
 
-.. release:: Upcoming
+.. release:: 1.8.0
+    :date: 2019-04-04
 
     .. change:: changed